bugs fixed and learned new things

+++ src/utils/cloudinary.js
@@ -1,29 +1,30 @@
 import { v2 as cloudinary } from 'cloudinary';
-import fs from "fs"
-import { url } from 'inspector';
+import fs from "fs";
 
- cloudinary.config({ 
-        cloud_name: 'process.env.CLOUDINARY_CLOUD_NAME', 
-        api_key: 'process.env.CLOUDINARY_API_KEY', 
-        api_secret: 'process.env.CLOUDINARY_API_SECRET' 
-    });
+cloudinary.config({
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET,
+});
 
-const uploadOnCloudinary = async(localFilePath) => {
+const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if(!localFilePath) return null
-        //upload the file on cloudnary
-        const respone = await cloudinary.uploader.upload
-        (localFilePath , {
-            resource_type : "auto"
-        })
-        // file has been uploaded successfully
-        console.log("file is uploaded on cloudinary" , respone.url); 
+        if (!localFilePath) return null;
+
+        const response = await cloudinary.uploader.upload(localFilePath, {
+            resource_type: "auto",
+        });
+
+        // console.log("file is uploaded on cloudinary", response.url);
+        fs.unlinkSync(localFilePath)
         return response;
-
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove locally saved temporily file as operation got failed
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+        console.error("Cloudinary upload failed:", error);
         return null;
-
-
     }
-}
+};
+
+export { uploadOnCloudinary };
